refactor: separate indexer into files per struct
