--- indexer_test.go
@@ -1,249 +1,0 @@
-package indexer
-
-import (
-	"errors"
-	"fmt"
-	"net/http"
-	"testing"
-
-	"github.com/jarcoal/httpmock"
-	"github.com/pokt-foundation/pocket-go/mock-client"
-	"github.com/pokt-foundation/pocket-go/provider"
-	testMock "github.com/stretchr/testify/mock"
-	"github.com/stretchr/testify/require"
-)
-
-type writerMock struct {
-	testMock.Mock
-}
-
-func (w *writerMock) WriteBlock(block *Block) error {
-	args := w.Called(block)
-
-	return args.Error(0)
-}
-
-func (w *writerMock) WriteTransactions(txs []*Transaction) error {
-	args := w.Called(txs)
-
-	return args.Error(0)
-}
-
-func (w *writerMock) WriteAccount(account *Account) error {
-	args := w.Called(account)
-
-	return args.Error(0)
-}
-
-func (w *writerMock) WriteNodes(nodes []*Node) error {
-	args := w.Called(nodes)
-
-	return args.Error(0)
-}
-
-func (w *writerMock) WriteApps(apps []*App) error {
-	args := w.Called(apps)
-
-	return args.Error(0)
-}
-
-func TestIndexer_IndexBlockTransactions(t *testing.T) {
-	c := require.New(t)
-
-	httpmock.Activate()
-	defer httpmock.DeactivateAndReset()
-
-	reqProvider := provider.NewProvider("https://dummy.com", []string{})
-
-	writerMock := &writerMock{}
-
-	indexer := NewIndexer(reqProvider, writerMock)
-
-	mock.AddMockedResponseFromFile(http.MethodPost, fmt.Sprintf("%s%s", "https://dummy.com", provider.QueryBlockTXsRoute),
-		http.StatusInternalServerError, "samples/query_block_txs.json")
-
-	err := indexer.IndexBlockTransactions(30363)
-	c.Equal(provider.Err5xxOnConnection, err)
-
-	mock.AddMockedResponseFromFile(http.MethodPost, fmt.Sprintf("%s%s", "https://dummy.com", provider.QueryBlockTXsRoute),
-		http.StatusOK, "samples/query_block_txs_empty.json")
-
-	err = indexer.IndexBlockTransactions(30363)
-	c.Equal(ErrNoTransactionsToIndex, err)
-
-	mock.AddMultipleMockedResponses(http.MethodPost, fmt.Sprintf("%s%s", "https://dummy.com", provider.QueryBlockTXsRoute),
-		http.StatusOK, []string{
-			"samples/query_block_txs.json",
-			"samples/query_block_txs_empty.json",
-			"samples/query_block_txs.json",
-			"samples/query_block_txs_empty.json",
-		})
-
-	writerMock.On("WriteTransactions", testMock.Anything).Return(errors.New("forced failure")).Once()
-
-	err = indexer.IndexBlockTransactions(30363)
-	c.EqualError(err, "forced failure")
-
-	writerMock.On("WriteTransactions", testMock.Anything).Return(nil).Once()
-
-	err = indexer.IndexBlockTransactions(30363)
-	c.NoError(err)
-}
-
-func TestIndexer_IndexBlock(t *testing.T) {
-	c := require.New(t)
-
-	httpmock.Activate()
-	defer httpmock.DeactivateAndReset()
-
-	reqProvider := provider.NewProvider("https://dummy.com", []string{})
-
-	writerMock := &writerMock{}
-
-	indexer := NewIndexer(reqProvider, writerMock)
-
-	mock.AddMockedResponseFromFile(http.MethodPost, fmt.Sprintf("%s%s", "https://dummy.com", provider.QueryBlockRoute),
-		http.StatusInternalServerError, "samples/query_block.json")
-
-	err := indexer.IndexBlock(30363)
-	c.Equal(provider.Err5xxOnConnection, err)
-
-	mock.AddMockedResponseFromFile(http.MethodPost, fmt.Sprintf("%s%s", "https://dummy.com", provider.QueryBlockRoute),
-		http.StatusOK, "samples/query_block_empty.json")
-
-	err = indexer.IndexBlock(30363)
-	c.Equal(ErrBlockHasNoHash, err)
-
-	mock.AddMockedResponseFromFile(http.MethodPost, fmt.Sprintf("%s%s", "https://dummy.com", provider.QueryBlockRoute),
-		http.StatusOK, "samples/query_block.json")
-
-	writerMock.On("WriteBlock", testMock.Anything).Return(errors.New("forced failure")).Once()
-
-	err = indexer.IndexBlock(30363)
-	c.EqualError(err, "forced failure")
-
-	writerMock.On("WriteBlock", testMock.Anything).Return(nil).Once()
-
-	err = indexer.IndexBlock(30363)
-	c.NoError(err)
-}
-
-func TestIndexer_IndexAccount(t *testing.T) {
-	c := require.New(t)
-
-	httpmock.Activate()
-	defer httpmock.DeactivateAndReset()
-
-	reqProvider := provider.NewProvider("https://dummy.com", []string{})
-
-	writerMock := &writerMock{}
-
-	indexer := NewIndexer(reqProvider, writerMock)
-
-	mock.AddMockedResponseFromFile(http.MethodPost, fmt.Sprintf("%s%s", "https://dummy.com", provider.QueryAccountRoute),
-		http.StatusInternalServerError, "samples/query_account.json")
-
-	err := indexer.IndexAccount("ABCD", 30363)
-	c.Equal(provider.Err5xxOnConnection, err)
-
-	mock.AddMockedResponseFromFile(http.MethodPost, fmt.Sprintf("%s%s", "https://dummy.com", provider.QueryAccountRoute),
-		http.StatusOK, "samples/query_account.json")
-
-	writerMock.On("WriteAccount", testMock.Anything).Return(errors.New("forced failure")).Once()
-
-	err = indexer.IndexAccount("ABCD", 30363)
-	c.EqualError(err, "forced failure")
-
-	writerMock.On("WriteAccount", testMock.Anything).Return(nil).Once()
-
-	err = indexer.IndexAccount("ABCD", 30363)
-	c.NoError(err)
-}
-
-func TestIndexer_IndexBlockNodes(t *testing.T) {
-	c := require.New(t)
-
-	httpmock.Activate()
-	defer httpmock.DeactivateAndReset()
-
-	reqProvider := provider.NewProvider("https://dummy.com", []string{})
-
-	writerMock := &writerMock{}
-
-	indexer := NewIndexer(reqProvider, writerMock)
-
-	mock.AddMockedResponseFromFile(http.MethodPost, fmt.Sprintf("%s%s", "https://dummy.com", provider.QueryNodesRoute),
-		http.StatusInternalServerError, "samples/query_nodes.json")
-
-	addresses, err := indexer.IndexBlockNodes(30363)
-	c.Equal(provider.Err5xxOnConnection, err)
-	c.Empty(addresses)
-
-	mock.AddMockedResponseFromFile(http.MethodPost, fmt.Sprintf("%s%s", "https://dummy.com", provider.QueryNodesRoute),
-		http.StatusOK, "samples/query_nodes_empty.json")
-
-	addresses, err = indexer.IndexBlockNodes(30363)
-	c.Equal(ErrNoNodesToIndex, err)
-	c.Empty(addresses)
-
-	mock.AddMockedResponseFromFile(http.MethodPost, fmt.Sprintf("%s%s", "https://dummy.com", provider.QueryNodesRoute),
-		http.StatusOK, "samples/query_nodes.json")
-
-	writerMock.On("WriteNodes", testMock.Anything).Return(errors.New("forced failure")).Once()
-
-	addresses, err = indexer.IndexBlockNodes(30363)
-	c.EqualError(err, "forced failure")
-	c.Len(addresses, 1)
-	c.Equal("98a18a38aa6826a55dccce19f607e3171cf1436e", addresses[0])
-
-	writerMock.On("WriteNodes", testMock.Anything).Return(nil).Once()
-
-	addresses, err = indexer.IndexBlockNodes(30363)
-	c.NoError(err)
-	c.Len(addresses, 1)
-	c.Equal("98a18a38aa6826a55dccce19f607e3171cf1436e", addresses[0])
-}
-
-func TestIndexer_IndexBlockApps(t *testing.T) {
-	c := require.New(t)
-
-	httpmock.Activate()
-	defer httpmock.DeactivateAndReset()
-
-	reqProvider := provider.NewProvider("https://dummy.com", []string{})
-
-	writerMock := &writerMock{}
-
-	indexer := NewIndexer(reqProvider, writerMock)
-
-	mock.AddMockedResponseFromFile(http.MethodPost, fmt.Sprintf("%s%s", "https://dummy.com", provider.QueryAppsRoute),
-		http.StatusInternalServerError, "samples/query_apps.json")
-
-	addresses, err := indexer.IndexBlockApps(30363)
-	c.Equal(provider.Err5xxOnConnection, err)
-	c.Empty(addresses)
-
-	mock.AddMockedResponseFromFile(http.MethodPost, fmt.Sprintf("%s%s", "https://dummy.com", provider.QueryAppsRoute),
-		http.StatusOK, "samples/query_apps_empty.json")
-
-	addresses, err = indexer.IndexBlockApps(30363)
-	c.Equal(ErrNoAppsToIndex, err)
-	c.Empty(addresses)
-
-	mock.AddMockedResponseFromFile(http.MethodPost, fmt.Sprintf("%s%s", "https://dummy.com", provider.QueryAppsRoute),
-		http.StatusOK, "samples/query_apps.json")
-
-	writerMock.On("WriteApps", testMock.Anything).Return(errors.New("forced failure")).Once()
-
-	addresses, err = indexer.IndexBlockApps(30363)
-	c.EqualError(err, "forced failure")
-	c.Len(addresses, 1)
-	c.Equal("98a18a38aa6826a55dccce19f607e3171cf14366", addresses[0])
-
-	writerMock.On("WriteApps", testMock.Anything).Return(nil).Once()
-
-	addresses, err = indexer.IndexBlockApps(30363)
-	c.NoError(err)
-	c.Len(addresses, 1)
-	c.Equal("98a18a38aa6826a55dccce19f607e3171cf14366", addresses[0])
-}
